# example/backend/output_backend/openapi_specs.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/measure/": {"get": {"summary": "Get All Measure", "operationId": "get_all_measure_measure__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}, "post": {"summary": "Create Measure", "operationId": "create_measure_measure__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MeasureCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/measure/{measure_id}/": {"get": {"summary": "Get Measure", "operationId": "get_measure_measure__measure_id___get", "parameters": [{"name": "measure_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Measure Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"summary": "Update Measure", "operationId": "update_measure_measure__measure_id___put", "parameters": [{"name": "measure_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Measure Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MeasureCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete Measure", "operationId": "delete_measure_measure__measure_id___delete", "parameters": [{"name": "measure_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Measure Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MeasureCreate": {"properties": {"value": {"type": "integer", "title": "Value"}, "day": {"type": "string", "format": "date", "title": "Day"}}, "type": "object", "required": ["value", "day"], "title": "MeasureCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}